fish_pipeline_v3/requirements.txt
torch
opencv-python
albumentations
scikit-learn
ultralytics
Pillow
numpy
matplotlib
